+++ CHANGELOG.md
@@ -65,6 +65,7 @@
 - The parameter `category_id` for the route `pim_enrich_product_listcategories` has been renamed to `categoryId`
 - Change constructor of `Pim\Bundle\BaseConnectorBundle\Reader\File\CsvProductReader`. Now `FieldNameBuilder`, channel, locale and currency entity classes are mandatory.
 - Add getAvailableLocaleCodes in AttributeInterface
+- AttributeTypeRegistry replaces AttributeTypeFactory, changed constructors for AttributeManager, ProductValueFormFactory, AddAttributeTypeRelatedFieldsSubscriber
 
 ## Bug fixes
 - PIM-3332: Fix incompatibility with overriden category due to usage of ParamConverter in ProductController